+++ CHANGELOG.md
@@ -4,6 +4,16 @@ GovReady-Q Release Notes
 In Development
 --------------
 
+Application changes:
+
+* Downloading output documents in 'markdown' format no longer round-trips through pandoc when the template was authored as Markdown.
+* Fix caching bug with output documents in multiple formats.
+* Small UI improvements.
+
+App authoring changes:
+
+* Templates no longer die but instead show inline error messages when an undefined context variable is accessed.
+
 Deployment changes:
 
 * Docker launch failed if some environment variables were not passed in.